backend/requirements.txt
fastapi
uvicorn
requests
httpx
python-dotenv
dedalus-labs

